+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "al-codeactions" extension will be documented in this
 
 ## 0.2.26
 
+- New code action: "Make variable global"
 - Support Jsonc format for AppSourceCop.json (Json with comments)
 
 ## 0.2.25